Update Remove-AzApplicationGatewaySslCertificate.md | Incomplete Task
The current documentation, indicates that using Remove-AzApplicationGatewaySllCertificate is the only step to remove the certificate. While this page does showcase the use of the command, the task to remove the certificate should be reflected in its entirety. Therefore the Set-AzApplicationGateway command is needed in line 30. 

Other edit: removed unneeded spacing on line 29.